refactor(calc): object 변경, 양수 조건 수정, 네이밍 룰 적용
